% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies.plot.R
\name{anomalies.plot.freq}
\alias{anomalies.plot.freq}
\title{Fonction qui fait les plots des fréquences d'anomalies}
\usage{
anomalies.plot.freq(Freq.ano.an, path)
}
\arguments{
\item{Freq.ano.an}{data.frame des fréquences des anomalies. Résultats de la fonction get.anomalies.freq.an()}

\item{path}{Chemin de sauvegarde des images}
}
\value{
Plots des fréquences des anomalies
}
\description{
Fonction qui fait les plots des fréquences d'anomalies
}
