% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies.plot.R
\name{anomalies.plot}
\alias{anomalies.plot}
\title{Fonction qui fait les plots d'anomalies}
\usage{
anomalies.plot(
  Anomalies.R,
  Anomalies.E,
  quantiles,
  year.min,
  year.max,
  duree.eval,
  Freq.ano,
  chi2,
  path
)
}
\arguments{
\item{Anomalies.R}{data.frame des anomalies de la période de référence}

\item{Anomalies.E}{data.frame des anomalies de la période d'évaluation}

\item{quantiles}{quantiles des anomalies. Résultats de la fonction get.quantile.ano()}

\item{year.min}{Numérique première année de référence}

\item{year.max}{Numérique dernière année d'évaluation}

\item{duree.eval}{Numérique durée de la période d'évaluation}

\item{Freq.ano}{Fréquence des anomalies. Résultats de la fonction get.anomalies.freq()}

\item{chi2}{Résultats du chi2}

\item{path}{Chemin de sauvegarde des images}
}
\value{
Plots des anomalies
}
\description{
Fonction qui fait les plots d'anomalies
}
