% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoread.R
\name{autoread}
\alias{autoread}
\title{Fonction qui lit automatique les fichiers en détectant tout}
\usage{
autoread(x, verbose = F, ...)
}
\arguments{
\item{x}{Un fichier csv}

\item{verbose}{Logique : écrire les messages ou non}
}
\value{
Plots des fréquences des anomalies
}
\description{
Fonction qui lit automatique les fichiers en détectant tout
}
