% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.taille.R
\name{Conversion.taille}
\alias{Conversion.taille}
\title{Fonction qui sort le tableau des tailles basé sur les pigments}
\usage{
Conversion.taille(data)
}
\arguments{
\item{data}{Tableau de données de pigments issu de PhytoCly}
}
\value{
Le tableau de données des fractions et biomasses CHLA issues des pigments
}
\description{
Fonction qui sort le tableau des tailles basé sur les pigments
}
