% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Dummy data}
\format{
## `data`
A data frame with 37,162 rows and 8 columns:
\describe{
  \item{Date}{Date in decimal format}
  \item{Site}{Sites expressed as letters}
  \item{Param}{Parameters expressed as LETTERS}
  \item{Val}{Value}
  \item{Latitude}{Latitude}
  \item{Longitude}{Longitude}
  \item{Month}{Month in interger}
  \item{Year}{Year in interger}
  ...
}
}
\source{
<no source, data randomly created>
}
\usage{
data
}
\description{
Dummy data with sites and parameters
Report ...
}
\keyword{datasets}
