% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoread.R
\name{detect_file_encoding}
\alias{detect_file_encoding}
\title{Fonction qui détecte l'encodage d'un fichier}
\usage{
detect_file_encoding(file_path, verbose = F)
}
\arguments{
\item{file_path}{Un fichier csv}

\item{verbose}{Logique : écrire les messages ou non}
}
\value{
Plots des fréquences des anomalies
}
\description{
Fonction qui détecte l'encodage d'un fichier
}
