% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.pval.r
\name{format.pval}
\alias{format.pval}
\title{Fonction qui format la pvalue pour les graphiques en < 0.05, < 0.01 et < 0.001}
\usage{
\method{format}{pval}(pval)
}
\arguments{
\item{pval}{Valeur ou vecteur de pvalues}
}
\value{
pvalues mises en forme
}
\description{
Fonction qui format la pvalue pour les graphiques en < 0.05, < 0.01 et < 0.001
}
