% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.quantile.ano.R
\name{get.anomalies.freq}
\alias{get.anomalies.freq}
\title{Fonction qui récupère les fréquences des anomalies}
\usage{
get.anomalies.freq(Anomalies.R, Anomalies.E, quantiles)
}
\arguments{
\item{Anomalies.R}{Anomalies de la prériode de référence}

\item{Anomalies.E}{Anomalies de la période d'évaluation}

\item{quantiles}{quantiles issus de get.quantile.ano}
}
\value{
Les fréquences d'anomalies
}
\description{
Fonction qui récupère les fréquences des anomalies
}
