% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.anomalies.R.r
\name{get.anomalies}
\alias{get.anomalies}
\title{Fonction qui fait les plots de mann kendall saisonniers}
\usage{
get.anomalies(data, year.max, duree.eval)
}
\arguments{
\item{data}{Tableau de données brutes}

\item{year.max}{Année max de la TS}

\item{duree.eval}{Durée de la période d'évaluation en année}
}
\value{
Les valeurs des anomalies
}
\description{
Fonction qui fait les plots de mann kendall saisonniers
}
