% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.buffer.dates.R
\name{get.buffer.dates}
\alias{get.buffer.dates}
\title{Fonction qui récupère les dates autours du mois ciblé pour la moyenne intégrée sur les jours juliens}
\usage{
get.buffer.dates(tm, ty, cl)
}
\arguments{
\item{tm}{Numérique : mois ciblé}

\item{ty}{Numérique : année ciblé}

\item{cl}{Un calendrier issu de la fonction get.calendar()}
}
\value{
Le tableau de données des fractions et biomasses CHLA issues des pigments
}
\description{
Fonction qui récupère les dates autours du mois ciblé pour la moyenne intégrée sur les jours juliens
}
