% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.calendar.R
\name{get.calendar}
\alias{get.calendar}
\title{Fonction qui crée un calendrier}
\usage{
get.calendar(from, to)
}
\arguments{
\item{from}{Numérique : année du début du calender au format yyyy}

\item{to}{Numérique : année du début du calender au format yyyy}
}
\value{
Le caldendrier
}
\description{
Fonction qui crée un calendrier
}
