% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.chi2.R
\name{get.chi2}
\alias{get.chi2}
\title{Fonction qui effectue un chi2 et exporte en un seul objet différents sorties}
\usage{
get.chi2(Freq.ano, simulate.p.value, B, ...)
}
\arguments{
\item{Freq.ano}{Le tableau de contingence}

\item{simulate.p.value}{cf fonction chisq.test}

\item{B}{: cf fonction chisq.test}
}
\value{
Le résultats du chi2
}
\description{
Fonction qui effectue un chi2 et exporte en un seul objet différents sorties
}
