% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.df.eval.ref.R
\name{get.df.eval.ref}
\alias{get.df.eval.ref}
\title{Fonction qui crée les dataframe évaluation et référence}
\usage{
get.df.eval.ref(Data_TS_GF, GF1, GF2)
}
\arguments{
\item{Data_TS_GF}{Le tableau de données avec les groupes fonctionnels}

\item{GF1}{Texte groupe fonctionnel 1}

\item{GF2}{Texte groupe fonctionnel 2}
}
\value{
Le dataframe avec les groupes fonctionnels
}
\description{
Fonction qui crée les dataframe évaluation et référence
}
