% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.mks.R
\name{get.mks}
\alias{get.mks}
\title{Fonction qui récupère les résultats du mann kendall saisonnier}
\usage{
get.mks(Data_TS_ok)
}
\arguments{
\item{Data_TS_ok}{Le tableau de données à utiliser}
}
\value{
Les résultats du mann kendall saisonnier
}
\description{
Fonction qui récupère les résultats du mann kendall saisonnier
}
