% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{get.PI.hull}
\alias{get.PI.hull}
\title{Fonction qui calcul les enveloppes}
\usage{
get.PI.hull(xR, yR, p, outer = TRUE, inner = TRUE)
}
\arguments{
\item{xR}{Numérique : vecteur de coordonnées du groupe fonctionnel en x lors de la période de référence.}

\item{yR}{Numérique : vecteur de coordonnées du groupe fonctionnel en y lors de la période de référence.}

\item{p}{Numérique : la valeur de p é considérer pour évaluer les enveloppes.}

\item{outer}{Logique : faut-il considérer l'enveloppe externe. Par défaut TRUE.}

\item{inner}{Logique : faut-il considérer l'enveloppe interne. Par défaut TRUE.}
}
\value{
les enveloppes du PI
}
\description{
Fonction qui calcul les enveloppes
}
