% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_png.R
\name{gg_png}
\alias{gg_png}
\title{Fonction qui sauve un gpglot}
\usage{
gg_png(
  GG_plot,
  filename,
  width = 30,
  height = 20,
  units = "cm",
  res = 300,
  path = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{GG_plot}{Le ggplot}

\item{filename}{Le nom sous lequel le sauver}

\item{width}{La largeur de l'image. Par défaut 30cm}

\item{height}{La hauteur de l'image. Par défaut 20cm}

\item{units}{L'unité des hauteur et largeur. Par défaut cm}

\item{res}{La résolution de l'image}

\item{path}{Le chemin de sauvegarde}

\item{verbose}{Ecrire les sorties images ou non}
}
\value{
Un plot sauvegardé au chemin spécifié
}
\description{
Fonction qui sauve un gpglot
}
