% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.PI.R
\name{global.PI}
\alias{global.PI}
\title{Fonction qui sort le tableau des tailles basé sur les pigments}
\usage{
global.PI(Data_TS_ok, GF1, GF2, duree.eval, path)
}
\arguments{
\item{Data_TS_ok}{Tableau de données}

\item{GF1}{Groupe fonctionnel 1 (doit correspondre a un argument de la colonne Param)}

\item{GF2}{Groupe fonctionnel 2 (doit correspondre a un argument de la colonne Param)}
}
\value{
Le tableau de données des fractions et biomasses CHLA issues des pigments
}
\description{
Fonction qui sort le tableau des tailles basé sur les pigments
}
