% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.year.a.enlever.R
\name{id.year.enlever}
\alias{id.year.enlever}
\title{Fonction qui identifie les années/mois pour lesquelles l'évaluation n'est pas possible}
\usage{
id.year.enlever(Data_TS_month, nb.mois.min, nb.cores = detectCores()/2)
}
\arguments{
\item{Data_TS_month}{Les données}
}
\value{
Un dataframe avec les années et mois a enlever
}
\description{
Fonction qui identifie les années/mois pour lesquelles l'évaluation n'est pas possible
}
