% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpo.R
\name{interpo.plot}
\alias{interpo.plot}
\title{Fonction qui fait un plot avec les données interpolées}
\usage{
interpo.plot(Data_TS_ok, name = NA)
}
\arguments{
\item{Data_TS_ok}{Les données}
}
\value{
Les données avec colonnes de valeurs interpolées
}
\description{
Fonction qui fait un plot avec les données interpolées
}
