% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpo.R
\name{interpo}
\alias{interpo}
\title{Sript parent de l'interpolation des valeurs manquantes}
\usage{
interpo(Data_TS_month_year_ok)
}
\arguments{
\item{Data_TS_month_year_ok}{Les données}
}
\value{
Les données avec colonnes de valeurs interpolées
}
\description{
Sript parent de l'interpolation des valeurs manquantes
}
