% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipak.R
\name{Load.Install.pkg}
\alias{Load.Install.pkg}
\title{Fonction qui installe et charge les packages (CRAN et GitHub)}
\usage{
Load.Install.pkg(github.pkg, cran.pkg = NULL, upgrade = TRUE)
}
\arguments{
\item{pkg}{Vecteur avec nom de packages en texte}
}
\description{
Fonction qui installe et charge les packages (CRAN et GitHub)
}
