% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map.ggplot.R
\name{Map.ggplot}
\alias{Map.ggplot}
\title{Fonction qui génère un fond de carte avec ggplot}
\usage{
Map.ggplot(
  xlim = c(-15, 15),
  ylim = c(30, 60),
  proj.to = 2154,
  already.proj = FALSE,
  ...
)
}
\arguments{
\item{xlim}{Coordonnées gps en x}

\item{ylim}{Coordonnées gps en y}

\item{proj.to}{Numerique : epsg du système vers lequel projeter}

\item{already.proj}{Logical : les données sont-elles deja au systeme vers lequel projeter}
}
\value{
Un fond de carte
}
\description{
Fonction qui génère un fond de carte avec ggplot
}
