% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply.windows.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{A script to implement a hackish version of parallel:mclapply() on Windows machines.}
\usage{
mclapply.hack(...)
}
\description{
A script to implement a hackish version of parallel:mclapply() on Windows machines.
}
