% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/message_parallel.R
\name{message_parallel}
\alias{message_parallel}
\title{Function which prints a message using shell echo; useful for printing messages from inside mclapply when running in Rstudio}
\usage{
message_parallel(...)
}
\description{
Function which prints a message using shell echo; useful for printing messages from inside mclapply when running in Rstudio
}
