% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mksm.misc.R
\name{mksm.fill.Sen.df}
\alias{mksm.fill.Sen.df}
\title{Fonction qui créer le tableau Sendf pour les plots}
\usage{
mksm.fill.Sen.df(data.x, Sen_df, year.min, year.max, Month = TRUE)
}
\arguments{
\item{data.x}{Data issues de get.mks()}

\item{Sen_df}{dataframe Sendf}

\item{year.min}{Numérique année min}

\item{year.max}{Numérique année max}
}
\value{
dataframe Sendf
}
\description{
Fonction qui créer le tableau Sendf pour les plots
}
