% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mksm.misc.R
\name{mksm.get.seasonal.estimate}
\alias{mksm.get.seasonal.estimate}
\title{Fonction qui créer le tableau Sendf pour les plots}
\usage{
mksm.get.seasonal.estimate(mksm)
}
\arguments{
\item{mksm}{Le tableau de contingence}
}
\value{
dataframe Sendf
}
\description{
Fonction qui créer le tableau Sendf pour les plots
}
