% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mois.text.r
\name{Mois.text}
\alias{Mois.text}
\title{Fonction qui change les mois décimaux en mois en texte}
\usage{
Mois.text(x)
}
\arguments{
\item{x}{Un vecteur numérique de mois}
}
\value{
Les données phyto ok
}
\description{
Fonction qui change les mois décimaux en mois en texte
}
