% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Moyenne_integree.R
\name{Moyenne_integree_profondeur}
\alias{Moyenne_integree_profondeur}
\title{Fonction qui effectue une moyenne intégrée sur la profondeur}
\usage{
Moyenne_integree_profondeur(x, by, from, to)
}
\arguments{
\item{x}{Numérique : les valeurs à inrégrer}

\item{by}{Numérique : les profondeurs}

\item{from}{Numérique : profondeur mini à intégrer}

\item{to}{Numérique : profondeur maxi à intégrer}
}
\value{
valeurs moyennée
}
\description{
Fonction qui effectue une moyenne intégrée sur la profondeur
}
