% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{PI.calc}
\alias{PI.calc}
\title{Fonction qui calcul le PI}
\usage{
PI.calc(xE, yE, PI.hull, p)
}
\arguments{
\item{xE}{Numérique : vecteur de coordonnées du groupe fonctionnel en x lors de la période d'évaluation.}

\item{yE}{Numérique : vecteur de coordonnées du groupe fonctionnel en y lors de la période d'évaluation.}

\item{PI.hull}{Résultats de la fonction get.PI.hull().}

\item{p}{Numérique : la valeur de p é considérer pour évaluer les enveloppes.}
}
\value{
Les valeurs du PI
}
\description{
Fonction qui calcul le PI
}
