% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI.R
\name{PI.plot}
\alias{PI.plot}
\title{Fonction qui fait les plots du PI}
\usage{
PI.plot(dataR, dataE, col, ds = NULL, PI.hull, PI, path = OUT, GF1, GF2)
}
\arguments{
\item{dataR}{data.frame : Données de la période de référence}

\item{dataE}{data.frame : Données de la période d'évaluation}

\item{ds}{Origine de dataset valeurs entre Rephy, PhytoCly, Boussole}

\item{PI.hull}{Résultats de la fonction get.PI.hull().}

\item{PI}{Résultats de la fonction get.PI.calc().}

\item{path}{Chemin de sauvegarde des images}

\item{GF1}{Groupe fonctionnel 1}

\item{GF2}{Groupe fonctionnel 2}

\item{cols}{Sur quelle colonne effectué les couleurs. Valeurs entre Year et Month}
}
\value{
les plots du PI
}
\description{
Fonction qui fait les plots du PI
}
