% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Boussole.R
\name{Read.Boussole}
\alias{Read.Boussole}
\title{Fonction qui lie et charge les données Boussoles}
\usage{
Read.Boussole(xlsx, sheet, year.max)
}
\arguments{
\item{xlsx}{le chemin vers les données}

\item{sheet}{Le numéro de la feuille avec les données}

\item{year.max}{Année max a considérer}
}
\value{
Les données boussole
}
\description{
Fonction qui lie et charge les données Boussoles
}
