% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.Pigments.R
\name{Read.Pigments}
\alias{Read.Pigments}
\title{Fonction qui lit et charge les données PhytoCly}
\usage{
Read.Pigments(xlsx, nsheet, csv)
}
\arguments{
\item{xlsx}{Le chemin vers les données}

\item{nsheet}{Le nombre de feuilles à prendre}
}
\value{
Les données PhytoCly
}
\description{
Fonction qui lit et charge les données PhytoCly
}
