% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.rephy.R
\name{Read.rephy}
\alias{Read.rephy}
\title{Fonction qui groupes les tableaux}
\usage{
Read.rephy(data_rephy, data_phytobs, CodeTaxon, TF)
}
\arguments{
\item{data_rephy}{Données rephy}

\item{data_phytobs}{Données phytobs}

\item{CodeTaxon}{Tableau de regroupement des UT}

\item{TF}{Base fonctionnelle}
}
\value{
Les données phyto ok
}
\description{
Fonction qui groupes les tableaux
}
