% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read.zoo.R
\name{Read.zoo}
\alias{Read.zoo}
\title{Fonction qui groupes les tableaux}
\usage{
Read.zoo(data_ecotaxa, CodesTaxons, TF)
}
\arguments{
\item{TF}{Base fonctionnelle}

\item{data_rephy}{Données rephy}

\item{data_phytobs}{Données phytobs}

\item{CodeTaxon}{Tableau de regroupement des UT}
}
\value{
Les données phyto ok
}
\description{
Fonction qui groupes les tableaux
}
