% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Remove.duplicates.r
\name{Remove.duplicates}
\alias{Remove.duplicates}
\title{Fonction pour enlever les duplicats}
\usage{
Remove.duplicates(data, nb.cores = detectCores()/2)
}
\arguments{
\item{data}{Les données}

\item{nb.cores}{Numérique : Nombre de coeurs sur lesquels paralléliser}
}
\value{
Les données ok
}
\description{
Fonction pour enlever les duplicats
}
