% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id.year.a.enlever.R
\name{remove.years}
\alias{remove.years}
\title{Fonction qui enlève les années/mois pour lesquelles l'évaluation n'est pas possible}
\usage{
remove.years(
  Data_TS_month,
  Year.a.enlever,
  prop.annee.mini,
  nb.cores = detectCores()/2
)
}
\arguments{
\item{Data_TS_month}{Les données}

\item{Year.a.enlever}{Dataframe issu de id.year.enlever()}

\item{prop.annee.mini}{proportion d'année mini pour la période d'évaluation}

\item{nb.cores}{Nombre de coeurs pour la parallélisation}
}
\value{
Les données avec les années / mois en NA
}
\description{
Fonction qui enlève les années/mois pour lesquelles l'évaluation n'est pas possible
}
