% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_rows.R
\name{replicate_rows}
\alias{replicate_rows}
\title{Modification de la fonction du package threadr pour enlever le each obligatoire}
\usage{
replicate_rows(df, n, reset = TRUE, each = FALSE)
}
\arguments{
\item{df}{cf threadr::replicate_rows()}

\item{n}{cf threadr::replicate_rows()}

\item{reset}{cf threadr::replicate_rows()}

\item{each}{Logique : considérer each ou pas}
}
\value{
Le résultats du chi2
}
\description{
Modification de la fonction du package threadr pour enlever le each obligatoire
}
