% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.agreg.R
\name{TS.agreg}
\alias{TS.agreg}
\title{Fonction mère de l'agrégation temporelles par moyenne normale}
\usage{
TS.agreg(Data_TS, agg.func, year.min, year.max, season)
}
\arguments{
\item{Data_TS}{Les données}

\item{agg.func}{Texte : fonction pour aggréger}

\item{year.min}{Numérique : Année min}

\item{year.max}{Numérique : Année max}

\item{season}{Numérique : Numéro de saisons. Par défaut 1:12}
}
\value{
Les données agrégées
}
\description{
Fonction mère de l'agrégation temporelles par moyenne normale
}
